Various core work; other systems impacted.
Fixed bugs in the Controller, particularly one causing SQL syntax errors with WHERE clauses.
Redoing the Data Model.
Redoing the Data Controller (but retaining the old one till the new one is done.)
Work on the front end.
Work on the database schema.
Work on the database sample data (which is still out of date.)